Day 6

+++ MrTakuVetClinic/Services/PetService.cs
@@ -1,8 +1,12 @@
 ﻿using MrTakuVetClinic.DTOs.Pet;
+using MrTakuVetClinic.DTOs.User;
+using MrTakuVetClinic.DTOs.Visit;
 using MrTakuVetClinic.Entities;
 using MrTakuVetClinic.Interfaces;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace MrTakuVetClinic.Services
@@ -12,27 +16,93 @@ namespace MrTakuVetClinic.Services
         private readonly IPetRepository _petRepository;
         private readonly IPetTypeRepository _petTypeRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IVisitRepository _visitRepository;
 
-        public PetService(IPetRepository petRepository, IPetTypeRepository petTypeRepository, IUserRepository userRepository)
+        public PetService(IPetRepository petRepository, IPetTypeRepository petTypeRepository, IUserRepository userRepository, IVisitRepository visitRepository)
         {
             _petRepository = petRepository;
             _petTypeRepository = petTypeRepository;
             _userRepository = userRepository;
+            _visitRepository = visitRepository;
         }
 
-        public async Task<IEnumerable<Pet>> GetAllPetsAsync()
+        public async Task<IEnumerable<PetDto>> GetAllPetsAsync()
         {
-            return await _petRepository.GetAllAsync();
+            var pets = await _petRepository.GetAllPetsAsync();
+            return pets.Select(p => new PetDto { 
+                PetId = p.PetId,
+                PetTypeId = p.PetTypeId,
+                PetName = p.PetName,
+                Breed = p.Breed,
+                BirthDate = p.BirthDate,
+                User = new UserDto {
+                    FirstName = p.User.FirstName,
+                    MiddleName = p.User.MiddleName,
+                    LastName = p.User.LastName,
+                    Email = p.User.Email,
+                    Username = p.User.Username,
+                    Active = p.User.Active,
+                    UserType = p.User.UserType.TypeName
+                }
+            }).ToList();
         }
 
-        public async Task<Pet> GetPetByIdAsync(int id)
+        public async Task<PetDto> GetPetByIdAsync(int id)
         {
             var pet = await _petRepository.GetPetByIdAsync(id);
             if (pet == null)
             {
                 throw new Exception("Pet not found.");
             }
-            return pet;
+
+            return new PetDto {
+                PetId = pet.PetId,
+                PetTypeId = pet.PetTypeId,
+                PetName = pet.PetName,
+                Breed = pet.Breed,
+                BirthDate = pet.BirthDate,
+                User = new UserDto
+                {
+                    FirstName = pet.User.FirstName,
+                    MiddleName = pet.User.MiddleName,
+                    LastName = pet.User.LastName,
+                    Email = pet.User.Email,
+                    Username = pet.User.Username,
+                    Active = pet.User.Active,
+                    UserType = pet.User.UserType.TypeName
+                }
+            };
+        }
+
+        public async Task UpdatePetByIdAsync(int id, PetUpdateDto petUpdateDto)
+        {
+            var existingPet = await _petRepository.GetByIdAsync(id);
+            if (existingPet == null)
+            {
+                throw new ArgumentException("Pet record not found.");
+            }
+
+            if (petUpdateDto.PetName != null)
+            {
+                existingPet.PetName = petUpdateDto.PetName;
+            }
+
+            if (petUpdateDto.PetTypeId != null)
+            {
+                existingPet.PetTypeId = petUpdateDto.PetTypeId.Value;
+            }
+
+            if (petUpdateDto.Breed != null)
+            {
+                existingPet.Breed = petUpdateDto.Breed;
+            }
+
+            if (petUpdateDto.BirthDate != null)
+            {
+                existingPet.BirthDate = petUpdateDto.BirthDate;
+            }
+
+            await _petRepository.UpdateAsync(existingPet);
         }
 
         public async Task PostPetAsync(Pet pet)
@@ -45,8 +115,19 @@ namespace MrTakuVetClinic.Services
         }
 
         public async Task DeletePetAsync(int id)
-        {
-            // TODO: refactor & add validation
+        { 
+            var pet = await _petRepository.GetPetByIdAsync(id);
+            if (pet == null)
+            {
+                throw new ArgumentException("Pet record not found.");
+            }
+
+            var visits = await _visitRepository.GetAllVisitsAsync();
+            if (visits.FirstOrDefault(p => p.PetId == id) != null)
+            {
+                throw new ArgumentException("Cannot delete the pet record because it has associated visit records.");
+            }
+            
             await _petRepository.DeleteAsync(id);
         }
     }
